Renamed bloc subfolders
