+++ lib/blocs/fusion_bloc/fusion_bloc.dart
