+++ lib/blocs/fusion_bloc/fusion_event.dart
