+++ lib/blocs/fusion_bloc/fusion_state.dart
