+++ lib/blocs/map_bloc/map_bloc.dart
@@ -10,7 +10,7 @@ import '../../data/models/shuttle_stop.dart';
 import '../../data/models/shuttle_update.dart';
 import '../../data/repository/shuttle_repository.dart';
 import '../../global_widgets/shuttle_svg.dart';
-import '../on_tap/on_tap_bloc.dart';
+import '../on_tap_bloc/on_tap_bloc.dart';
 
 part 'map_event.dart';
 part 'map_state.dart';
