+++ lib/blocs/map_bloc/map_state.dart
