+++ lib/data/models/shuttle_stop.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_map/flutter_map.dart';
 
-import '../../blocs/on_tap/on_tap_bloc.dart';
+import '../../blocs/on_tap_bloc/on_tap_bloc.dart';
 import '../../global_widgets/stop.dart';
 import 'shuttle_point.dart';
