+++ lib/global_widgets/stop.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:latlong/latlong.dart';
 
-import '../blocs/on_tap/on_tap_bloc.dart';
+import '../blocs/on_tap_bloc/on_tap_bloc.dart';
 import '../pages/map_page/widgets/eta_panel.dart';
 
 class Stop extends StatelessWidget {
