+++ lib/pages/map_page/widgets/attribution.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import '../../../blocs/theme/theme_bloc.dart';
+import '../../../blocs/theme_bloc/theme_bloc.dart';
 import 'hyperlink.dart';
 
 class Attribution extends StatelessWidget {
