+++ lib/pages/map_page/widgets/legend.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
-import '../../../blocs/theme/theme_bloc.dart';
+import '../../../blocs/theme_bloc/theme_bloc.dart';
 import '../../../global_widgets/shuttle_svg.dart';
 import 'legend_row.dart';
