+++ lib/pages/routes_page/routes_page.dart
@@ -3,8 +3,8 @@ import 'dart:async';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
-import '../../blocs/routes/routes_bloc.dart';
-import '../../blocs/theme/theme_bloc.dart';
+import '../../blocs/routes_bloc/routes_bloc.dart';
+import '../../blocs/theme_bloc/theme_bloc.dart';
 import '../../global_widgets/loading_state.dart';
 import '../../global_widgets/shuttle_svg.dart';
 import 'states/loaded_routes.dart';
