+++ lib/pages/routes_page/widgets/panel.dart
@@ -4,8 +4,8 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:scrollable_positioned_list/scrollable_positioned_list.dart';
 import 'package:latlong/latlong.dart';
 
-import '../../../blocs/on_tap/on_tap_bloc.dart';
-import '../../../blocs/theme/theme_bloc.dart';
+import '../../../blocs/on_tap_bloc/on_tap_bloc.dart';
+import '../../../blocs/theme_bloc/theme_bloc.dart';
 import '../../../data/models/shuttle_stop.dart';
 import 'shuttle_line.dart';
