+++ lib/pages/schedules_page/schedules_page.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
-import '../../blocs/theme/theme_bloc.dart';
+import '../../blocs/theme_bloc/theme_bloc.dart';
 
 class SchedulesPage extends StatefulWidget {
   @override
