+++ lib/pages/settings_page/settings_page.dart
@@ -2,7 +2,7 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
-import '../../blocs/theme/theme_bloc.dart';
+import '../../blocs/theme_bloc/theme_bloc.dart';
 import 'widgets/android_settings.dart';
 
 class SettingsPage extends StatefulWidget {
