+++ lib/pages/settings_page/widgets/android_settings.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import '../../../blocs/theme/theme_bloc.dart';
+import '../../../blocs/theme_bloc/theme_bloc.dart';
 import 'about_settings.dart';
 import 'feedback_settings.dart';
 import 'general_settings.dart';
