+++ lib/pages/settings_page/widgets/ios_settings.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_cupertino_settings/flutter_cupertino_settings.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
-import '../../../blocs/theme/theme_bloc.dart';
+import '../../../blocs/theme_bloc/theme_bloc.dart';
 
 class IOSSettings extends StatefulWidget {
   final ThemeState theme;
